+++ packages/app/navigation/ToolsNavigator.tsx
@@ -11,6 +11,7 @@ import { DrawerNavigationProp } from "@react-navigation/drawer";
 const SearchScreen = lazy(() => import("../screens/Tools/Search"));
 const CalendarScreen = lazy(() => import("../screens/Tools/Calendar"));
 const CreditsScreen = lazy(() => import("../screens/Tools/Credits"));
+const TestScanScreen = lazy(() => import("../screens/Tools/TestScan"));
 const OpenSourceScreen = lazy(() => import("../screens/Tools/OpenSource"));
 const BouncersScreen = lazy(() => import("../screens/Tools/Bouncers"));
 const BouncersExpiringScreen = lazy(() => import("../screens/Tools/BouncersExpiring"));
@@ -24,6 +25,7 @@ const MunzeeScreen = lazy(() => import("../screens/Tools/Munzee"));
 const TypeCategoryScreen = lazy(() => import("../screens/Tools/Types/Category"));
 const TypeMunzeeScreen = lazy(() => import("../screens/Tools/Types/Type"));
 const ActivityWidgetScreen = lazy(() => import("../screens/Tools/WidgetConfigure/ActivityWidget"));
+const GardenPainterScreen = lazy(() => import("../screens/Tools/Garden/Painter"));
 
 const ToolsStack = createStackNavigator<ToolsStackParamList>();
 
@@ -42,21 +44,42 @@ export default function ToolsNavigator() {
         },
       }}>
       <ToolsStack.Screen name="Search" component={SearchScreen} />
+
       <ToolsStack.Screen name="Calendar" component={CalendarScreen} />
-      <ToolsStack.Screen name="Credits" component={CreditsScreen} />
-      <ToolsStack.Screen name="OpenSource" component={OpenSourceScreen} />
+      <ToolsStack.Screen name="EvoPlanner" component={EvoPlannerScreen} />
+      <ToolsStack.Screen name="TestScan" component={TestScanScreen} />
+      <ToolsStack.Screen name="WidgetConfigureActivityWidget" component={ActivityWidgetScreen} />
+
       <ToolsStack.Screen name="Bouncers" component={BouncersScreen} />
       <ToolsStack.Screen name="BouncersExpiring" component={BouncersExpiringScreen} />
-      <ToolsStack.Screen name="BouncersMap" component={BouncersMapScreen} />
       <ToolsStack.Screen name="Nearby" component={NearbyScreen} />
-      <ToolsStack.Screen name="Donate" component={DonateScreen} />
+      <ToolsStack.Screen name="BouncersMap" component={BouncersMapScreen} />
+
       <ToolsStack.Screen name="POIPlanner" component={POIPlannerScreen} />
       <ToolsStack.Screen name="DestinationPlanner" component={DestinationPlannerScreen} />
-      <ToolsStack.Screen name="EvoPlanner" component={EvoPlannerScreen} />
-      <ToolsStack.Screen getId={({params}) => `${params.a}/${params.b}`} name="Munzee" component={MunzeeScreen} />
-      <ToolsStack.Screen getId={({params}) => params.category} name="TypeCategory" component={TypeCategoryScreen} />
-      <ToolsStack.Screen getId={({params}) => params.type} name="TypeMunzee" component={TypeMunzeeScreen} />
-      <ToolsStack.Screen name="WidgetConfigureActivityWidget" component={ActivityWidgetScreen} />
+
+      <ToolsStack.Screen name="GardenPainter" component={GardenPainterScreen} />
+
+      <ToolsStack.Screen
+        getId={({ params }) => `${params.a}/${params.b}`}
+        name="Munzee"
+        component={MunzeeScreen}
+      />
+
+      <ToolsStack.Screen
+        getId={({ params }) => params.category}
+        name="TypeCategory"
+        component={TypeCategoryScreen}
+      />
+      <ToolsStack.Screen
+        getId={({ params }) => params.type}
+        name="TypeMunzee"
+        component={TypeMunzeeScreen}
+      />
+
+      <ToolsStack.Screen name="Credits" component={CreditsScreen} />
+      <ToolsStack.Screen name="Donate" component={DonateScreen} />
+      <ToolsStack.Screen name="OpenSource" component={OpenSourceScreen} />
     </ToolsStack.Navigator>
   );
 }