+++ packages/app/screens/Tools/Garden/Painter.tsx
@@ -4,12 +4,12 @@ import { useNavigation } from "@react-navigation/native";
 import { Button, Layout, Text } from "@ui-kitten/components";
 import * as React from "react";
 import { Pressable, View } from "react-native";
-import TypeImage from "../../components/Common/TypeImage";
-import MapView from "../../components/Maps/MapView";
-import useMunzeeRequest from "../../hooks/useMunzeeRequest";
-import useTitle from "../../hooks/useTitle";
+import TypeImage from "../../../components/Common/TypeImage";
+import MapView from "../../../components/Maps/MapView";
+import useMunzeeRequest from "../../../hooks/useMunzeeRequest";
+import useTitle from "../../../hooks/useTitle";
 import Clipboard from "expo-clipboard";
-import Icon from "../../components/Common/Icon";
+import Icon from "../../../components/Common/Icon";
 
 export default function BouncersMapScreen() {
   const nav = useNavigation();