+++ packages/functions/src/user/inventory.ts
@@ -1,5 +1,15 @@
+import fetch from "node-fetch";
 import { Route } from "../types";
 import { request } from "../util";
+
+async function getQrates(token: string) {
+  try {
+    const response = await fetch(`https://api.munzee.com/qrates?access_token=${encodeURIComponent(token)}`);
+    return (await response.json()).data;
+  } catch (e) {
+    return null;
+  }
+}
 
 const route: Route = {
   path: "user/inventory",
@@ -8,11 +18,12 @@ const route: Route = {
     {
       version: 1,
       async function({ params: { access_token } }) {
-        var [undeployed, credits, history, boosters] = await Promise.all([
+        var [undeployed, credits, history, boosters, qrates] = await Promise.all([
           request("user/undeploys/count", {}, access_token),
           request("user/credits", {}, access_token),
           request("user/credits/history", {}, access_token),
           request("user/boosters/credits", {}, access_token),
+          getQrates(access_token)
         ] as const);
         const formattedUndeployed = undeployed?.data?.map(i => ({
           type: i.pin_icon.match(/\/([^./]+).png/)?.[1],
@@ -25,6 +36,7 @@ const route: Route = {
             history: history?.data,
             boosters: boosters?.data,
             undeployed: formattedUndeployed,
+            qrates,
           },
         };
       },